instance-creation unit test: cloud-init.go tested
